% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ps_params}
\alias{ps_params}
\title{Pushshift.io parameters for all endpoints}
\format{
tibble of Endpoint Paramters
}
\source{
https://pushshift.io/api-parameters/
}
\usage{
ps_params
}
\description{
Pushshift.io parameters for all endpoints
}
\keyword{datasets}
