% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_comments}
\alias{search_comments}
\title{Search Pushshift.io for Reddit comments}
\usage{
search_comments(
  q = "",
  subreddit = NULL,
  size = 25,
  after = NULL,
  before = NULL,
  ...
)
}
\arguments{
\item{q}{character, Query term}

\item{subreddit}{Restrict results to subreddit (use "!" to negate, comma delimited for multiples)}

\item{size}{integer, number of results to return}

\item{after}{Restrict results to those made after this epoch time}

\item{before}{Restrict results to those made before this epoch time}

\item{...}{Further arguments passed as query parameters. See [ps_params] for all possible parameters.}
}
\value{
tibble of comments
}
\description{
Search Pushshift.io for Reddit comments
}
