% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_epoch}
\alias{to_epoch}
\title{convert date to epoch time}
\usage{
to_epoch(date)
}
\arguments{
\item{date}{date to be converted}
}
\value{
epoch time as integer
}
\description{
convert date to epoch time
}
\examples{
to_epoch("2022-12-01")
}
